+++ docs/changelog.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Upcoming
 
+### Added
+
+- Added param library to choose the icon library also in `six-icon-button` and `six-tile` (default
+  is material-icons)
+
 ## 5.3.1 - 2026-03-06
 
 ### Added